// rently-client/src/schemas/role.schema.ts
import { z } from "zod";

export const RoleSchema = z.object({
  id: z.number(),
  name: z.string().max(500, "Tên vai trò không hợp lệ"),
  description: z.string(),
  isActive: z.boolean().default(true),
  createdById: z.number().nullable(),
  updatedById: z.number().nullable(),
  deletedById: z.number().nullable(),
  deletedAt: z.date().nullable(),
  createdAt: z.date(),
  updatedAt: z.date(),
});

export type RoleType = z.infer<typeof RoleSchema>;